+++ packages/benefit-docs/gatsby-node.js
@@ -29,7 +29,7 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
       // (or `node.frontmatter.slug`)
       path: node.frontmatter.route,
       // This component will wrap our MDX content
-      component: path.resolve(`./src/components/docs-page-layout.js`),
+      component: path.resolve(`./src/components/Page.js`),
       // We can use the values in this context in
       // our page layout component
       context: { id: node.id },
